Tomsk_kedr/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-phonenumber-field==7.0.2
djangorestframework==3.16.1
phonenumbers==9.0.17
phonenumberslite==9.0.17
sqlparse==0.5.3
tzdata==2025.2
